Fix build issues when building on ARM 64 platforms
The first change changes the type of ch from char to int, since it needs
to hold negative values and char is unsigned on some machines.
The second change replaces the Intel x86 fence instructions with the
ARM 64 equivalents.
The last two changes remove the "march=-native" CFLAG and arrange to
have dpdk built with T=arm64-armv8a-linuxapp-gcc option.

Change-Id: Iea6d0587f704fadf091fd731355eb07911c22d5a
Signed-off-by: Barry Spinney <spinney@mellanox.com>