test/bdev: handle the case the IO channel is not setup
This issue may be hit under some large IO and large queue
depth cases where much memory will be consumed and the channel
may be failed to create due to insufficient memory.

The general fix here is that we need to properly handle
the error case that the channel is not created due to this
or that reason.

Change-Id: Icf58a7c8a9cba625ef1b8da6b8bf68ef72ef5546
Signed-off-by: GangCao <gang.cao@intel.com>