test/iscsi: do more cleanup on iscsi test failures
There are some tests which when failing, leave the
iSCSI initiator in a bad state - even after a reboot.
So do a sendtargets w/ delete to clean this up.

For now, do this just in the iSCSI filesystem test,
since this is the first iSCSI test to run.  Long-term
we need to look at where best to put this.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: Ibc0cd3c65e1f39ab6e0d27d2d6ca6fe74c12d477