bdev: add test unit ready before media access commands
This is not for final review, just looking to get input on the
correct way to do this. Ignore the extra debug prints that are in
here, they will be gone before it's ready for real review (and
test code will be added).

For this quick hack, what I did was:

- intercept media access commands (added new inline func to decide)
- hijack the incoming bdev to use its vreq to create the TUR
- save off the original bdev callback and caller_ctx in the io_ctx
as new fields
- replace with the TUR callback and the channel that's needed in
the TUR competion
- in the poller for virtio requests check to see if its an internal
req or not based on whether the new fields in the io_ctx are NULL or
not and if so call the cb directly instead of via bdev API because
the bdev_ch outstanding IO count will get out of whack if I complete
the internally generated TUR via bdev
- in the TUR callback restore the original bdev cb and caller_ctx,
NUL out the saved fields in the io_ctx and continue with the original
request (will send start/stop at this point based on resp but no
point in adding that until I understand if this general method is
correct or not)

There may be a waaaay better way to do this (generate intneral IO) but
I didn't see any easy to follow examples anywhere so feel free to tell
me this is totally wrong but please help me understand the right way
to allocate, submit and complete an IO from within a bdev module.

Change-Id: I79882952a41b083597a44525ff503e766fe441dc
Signed-off-by: Paul Luse <paul.e.luse@intel.com>