scsi: Make the pending task behavior correct for reset
Previously, our task has bdev_io pointer, and we can
set bdev_io status to make it not report to upperlayer.
But currently, we do not have this pointer, so the methods
is to we set task status to condition, but we did not return
direclty, we should still wait the bdev layer return the bdev
io status, then we trigger the abort commands.

Change-Id: I6af2bda0f0b1de7b0c655d2ac2980ddca48c1162
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>