Update patch set 2
Patch Set 2: Code-Review-1

Hi Ziye,

Could you take a look at the commit message again?  It is not very clear to me.

It would be good to explain with more specifics exactly what test fails without this patch.  Anything on pending_tasks list has not been submitted to the bdev yet, so it should be safe to abort immediately.  In this case we do not need to worry about the bdev layer completing the I/O later, since the I/O was never submitted to the bdev layer.

Shuhei - does my comment address your concern?  For this case in Ziye's patch, these I/O have not been submitted to bdev layer.

-Jim

Patch-set: 2
Label: Code-Review=-1