Update patch set 2
Patch Set 2: Code-Review-1

Hi Maciek,

Some devices support unmap, but subsequent reads to those sectors are not guaranteed to be zero.  So this patch would break blobstore on those devices - we *need* all of the metadata pages to be all zeroes when the blobstore is created.

For drives that do not support write zeroes, the bdev layer *should* be translating these to normal write commands with a zeroed buffer.  Is this not working?

-Jim

Patch-set: 2
Label: Code-Review=-1