Update patch set 2
Patch Set 2:

Hi Jim,
this works correctly, but there is a case from mailing list ("lvol function: hangs up with Nvme bdev.") where device does not support write zeros, so we do it just by writing it and it takes a lot of time to do this and create blob store. I did not know that unmap do not guarantee that. In that case, is there anything else that we can do to speed this up?

Patch-set: 2
Reviewer: Maciej Szwed <1011262@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0