scsi: Make the complete task behavior correct for reset
In principle, we should not have active tasks, which means
that lun->tasks should be empty, So we may need to remove
the loop.

If we want to keep this loop, we should not directly
call spdk_scsi_lun_complete_task since those tasks
are sent to bdev layer. And finally, those tasks will
be return (even aborted) by call backs.
So we only need to set task status to condition, but
not call spdk_scsi_lun_complete_task directly.

Change-Id: I6af2bda0f0b1de7b0c655d2ac2980ddca48c1162
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>