Update patch set 3
Patch Set 3:

See my new patch:

In my view is that there should not be active tasks, those should all be handled during reset. 

If we want to keep this loop(which means there are outing I/Os), we should not directly call spdk_scsi_lun_complete_task to return, this is not wrong.

Patch-set: 3
Label: Code-Review=0