Update patch set 2
Patch Set 2:

With assumption that some devices do not return zero on unmap. Blobstore init could write zeroes on super block and metadata region, and unmap rest of space on the device. This way condition of having 0's on metadata would satisfied and rest of device would be at least unmapped.

This keeps time to init blobstore minimal on devices that do not support write_zeroes but do support unmap.

Patch-set: 2
Label: Verified=0