Update patch set 3
Patch Set 3:

I'm not sure about this patch.  Assigning to Ben to get his input.  Exiting seems drastic but maybe that's the right thing to do.   At a minimum, we should add the error message - then it's just a matter of assert(false) v. exit(-1).

What if instead of exit(-1), we do assert(false) but then if we get an I/O and the qpair is NULL we just fail it immediately?

Patch-set: 3
Label: Verified=0