Update patch set 1
Patch Set 1:

The concern here is that the application did not exit by itself when the portal is not opened. Ctrlc+c can kill the application.

The fix here is to exit the application when all the portal are not opened. If it is ok to leave the application there, then this fix is not needed.

My thinking is that if we enable "RPC", then we can will not exit as RPC can create a new portal. If RPC is not enable, it's better to just exit as we may do nothing to make the iSCSI target responsive.

What's your thinking? Thanks.

Patch-set: 1
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0