Update patch set 1
Patch Set 1: Code-Review-1

My concern is that this consumes a lot of extra memory per connection.  This is 192 more PDUs per connection, for default 128 connections in the iSCSI target.  Each spdk_iscsi_pdu is 248 bytes.  192 * 248 * 128 = 6MB.  It's not a huge amount, but if we bump this up to 1024 or 2048 default number of connections in the future, it becomes much bigger.

We could add error handling when there are no PDUs available - this might be a nice improvement in the future.

What happens when we support RAID-0 of two P4500 devices?  Then 256 is not enough, we need 512.

I'm just not sure that trying to get max performance from one iSCSI session is realistic.  I would be surprised if one session can get 500K IO/s * 4KB/IO = 2GB/s of throughput - that is close to saturating a single 25Gb link.

Patch-set: 1
Label: Code-Review=-1