Update patch set 22
Patch Set 22: Code-Review-1

I think vhost.conf.in should just be removed.  This patch needs to be rebased on top of the recent changes that removed the need to enable RPC (its not enabled by default).  The only thing remaining is LogFacility which is no longer used or needed.

Patch-set: 22
Label: Code-Review=-1