Update patch set 3
Patch Set 3: Code-Review-1

Is this patch needed or just an optimization?

My concern is that we're adding an assumption here without any asserts that could ever save us if the assumption was ever broken.  Keeping the code as-is is safer.

Maybe we add a new spdk_poller_get_core() function so we can assert that we're stopping the poller on the poller's core?

Patch-set: 3