Update patch set 3
Patch Set 3: Code-Review-1

Can you describe the specific test case that causes the error fixed by this patch?

I just want to make sure we are not fixing something here but there is really a problem somewhere else that we need to look at.  Does the io_channel allocation really fail because of a memory allocation failure, or is it because there is some other resource not available (for example, run out of NVMe queues)?

Patch-set: 3
Label: Code-Review=-1