Update patch set 3
Patch Set 3: Code-Review-1

I think this is a step in the right direction, but we can simplify it even more.

We should be able to just allocate the tasks individually - we don't need the mempool at all if we are doing all of the allocations up front.

See the similar change in commit b163dc02168c72f8759f134e12d55dbe556745cc ("nvme/perf: replace DPDK uses with env abstractions").

Patch-set: 3
Label: Code-Review=-1