Update patch set 4
Patch Set 4:

I pushed the following change:
https://review.gerrithub.io/#/c/388334/

The purpose of this patch is to abort any task found in the lun->tasks list unexpectedly.

I think the object reuse mechanism based on the sequence number is helpful and can be applied to SPDK but I'm not sure whether this match SPDK's policy.

Thank you.

Patch-set: 4