Update patch set 1
Patch Set 1:

I do not think that we need to do it in this way.
In my mind, we should use the mechanism provided in https://review.gerrithub.io/#/c/386817/
If we find those active tasks, we just set the status since it will be always returned by the backend, and in scsi spdk_bdev_scsi_task_complete_cmd we can check the task status. If it is aborted, we directly abandom the status from the bottom device.

Patch-set: 1
Reviewer: Ziye Yang <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Verified=0