Update patch set 9
Patch Set 9: Code-Review-1

Let's hold off on this patch for now.  We still need to bottom out on the RFC.  There are also versioning changes we need to add - for example, flags in metadata for blobs to mark it as thin provisioned, and how earlier versions of blobstore apps will handle those flags that it does not understand.

Patch-set: 9
Label: Code-Review=-1