Update patch set 9
Patch Set 9:

Hi Shuhei,

If the LUN cannot abort all I/O for LUN RESET, should we still return the LUN RESET with success status?

For tasks that still remain after reset - do we need to mark them as ABORTED?  Or if they complete later, is it OK to complete them with SUCCESS status?

-Jim

Patch-set: 9