Update patch set 9
Patch Set 9: Code-Review-1

Hi Jim, 

If the LUN cannot abort all I/O for LUN RESET, we should return the LUN RESET with failure status.
For tasks that still remain after reset - we need to mark them as ABORTED
Ziye’s patch satisfies both. 

I understood that the change 388548 and 386837 by Cunyin are for the following sequence:
- Initiator detected timeout for an IO
- Initiator issued a ABORT TASK for the IO.
- Target returned failure to Initiator because ABORT TASK is not supported.
- Initiator’ error handling was promoted and Initiator issued LUN RESET.
- Target handled LUN RESET but took a long time to complete.
- Initiator detected timeout for LUN RESET.
- Initiator issued an new login request.

Both of LUN RESET and new login request were the result of promotion of the error handling and LUN RESET was in the middle of the error handling. If LUN RESET fails, stronger command will be issued.

I think this change does not degrade performance and hence we should be conservative and should not return success for any uncontrollable IO.

Thanks, Shuhei

Patch-set: 9