Update patch set 3
Patch Set 3:

The patch itself is to handle the error condition at the high layer for the unallocated channel and when this situation does happen (means the failure of allocation channel), there needs some error recovery here.

What we are talking in previous comments is why the failure of allocation channel happened including the patch Ziye mentioned to make some changes on the allocation. This is the reason of the failure to allocate channel. However the fix is to handle what to do after the failure occurs.

Thanks,
Gang

Patch-set: 3
Reviewer: GangCao <1011207@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0