Update patch set 8
Patch Set 8:

I originally thought this might be the cause of issues with fio_plugin. It is apparently not and this patch is not `necessary` anymore. However it's possibly still a bug or at least weird behavior that can be easily changed.

> Can you walk me through an exact sequence of steps that was causing an issue?

Sure, I don't have the real stacktrace from a crash right now but the following is a reconstruction of functions called in proper order. The _spdk_bs_load_super_cpl is where the problem starts.

* spdk_bdev_register
  * vbdev_lvs_examine
    * spdk_lvs_load (with _vbdev_lvs_examine_cb finish callback)
      * spdk_bs_load (this creates io_device and a single io_channel)
        *spdk_bs_sequence_read (super block is read)
... (the following is called asynchronously)
* _spdk_bs_load_super_cpl
  * Version mismatch! (or any other error inside _spdk_bs_load_super_cpl). 
    * _spdk_bs_free is called
      * md_channel is put (but this is an async operation!)
      * md_target io_device is unregistered (which gets deferred due to an io_channel still present)
    * the sequence is finished
      * the lvol examine callback is called (_vbdev_lvs_examine_cb)
        * spdk_bdev_module_examine_done
          * spdk_bdev_init_complete
... (async, possibly in a next reactor cycle)
* the md_channel is really put and this calls the md_target_ io_device unregister callback, which destroys a blob & blobstore

Patch-set: 8
Reviewer: Dariusz Stojaczyk <1011218@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=0