Update patch set 18
Patch Set 18:

Ben and I talked over IRC and came up with a better and simpler solution.

When you get the reset command, do the for_each_channel immediately.  When it's done, if there were no I/O outstanding, complete the reset.  Otherwise, start a poller/timer for 1 millisecond.  When that expires do the for_each_channel again - same logic, if no I/O outstanding, complete the reset otherwise start a poller/time for 1 millisecond.

This will avoid races when there are multiple channels and will also ensure we always complete the reset on the same thread that submitted it.

Patch-set: 18