Update patch set 3
Patch Set 3: Code-Review-1

Hi Maciek,

This doesn't seem to match any of the patches I suggested in my e-mail yesterday.  Let's hold off on this one until we get some of the other patches done.  For example, I see thin provisioning as a feature bit in the new flags descriptor - not as an xattr.

-Jim

Patch-set: 3
Label: Code-Review=-1