Update patch set 2
Patch Set 2: Code-Review-1

This is not inline with the RFC we have been discussing.  We do not want explicit memsets in the I/O path here - it should be a generic read to a bs_dev.  For thin provisioned blobs without a backing store, the read would do the memset and complete the request immediately.

Please hold off on further work on this patch - let's tackle the patches I suggested in my e-mail first.

-Jim

Patch-set: 2
Label: Code-Review=-1