nvmf/tgt: fix issues for ctrlr+ c handling.
While receving ctrlr+c event, NVMe-oF target
could in any state. So we cannot guarantee
g_acceptor_poller is initialized. If
we do not handle such case, ctrlr+c will
trigger unexpected coredump issue.

Another case is that: if spdk_nvmf_parse_conf fails,
then g_acceptor_poller is also not initialized, so
the solution is that we add a new SHUTDOWN state.

Change-Id: Id89a96b5d39f8a528e72dea8c0eb6524bdaf7ee4
Signed-off-by: Ziye Yang <optimistyzy@gmail.com>