Update patch set 11
Patch Set 11:

I found and fixed the bug which caused discovery failure during the auto test. I explain the bug in the following.

For the SendTargets command, the initiator set All to request that all relevant targets are returned. SendTargets is not affected by the Initiator Group but I changed this All to Any by mistake.

I wanted to use Any only for the filter function of the Initiator Group.

By discarding the change in the SendTargets command, I could confirm the test passed.

I hope these make sense to you and please rerun the auto test.

Thanks

Patch-set: 11
Label: Code-Review=0