Update patch set 2
Patch Set 2:

This is actually the situation that you described: "For thin provisioned blobs without a backing store, the read would do the memset and complete the request immediately"
It's thin provisioned blob that is not a clone of snapshot. For clone the logic would be as you designed it in RFC - call to backing bs_dev. This case still needs to be implemented.

Patch-set: 2
Label: Code-Review=0