bdev/virtio: reuse io_channels for target scan
Rather than acquiring an exclusive virtqueue
for target scan, reuse an existing one by
temporarily switching it's poller to handle
scan I/O as well.

The goal is to be able to rescan the virtio
device in runtime (hot-attach/detach).

Now that scan is done on existing queues,
it is no longer guaranteed that rings have
enough space to enqueue all requests. This
will be fixed in subsequent patches.

Change-Id: I0a39460c0ddb52592ae6828e5a7f51056ada2696
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>