Update patch set 1
Patch Set 1: Code-Review+2

Sorry, I just looked at the next patch in the series, and it is changing the destruct path to remove the bdev from the list.

I think we can go ahead and merge both of these patches (or squash them together if we don't get to this one right away).

Patch-set: 1
Label: Code-Review=+2