[WIP] virtio: make virtio_req->req and resp optional
We assumed that each request contains
a single request and a single response
descriptor. However that's not the case
for virtio scsi eventq, where each event
contains only a response.

This patch makes the request and reponse
fields optional, but I'd say that's an
ugly hack (at least the way it's done now).
The virtio_req will probably need some
refactor...

Change-Id: Ie56dac3745cc7199d36409c546701792aa0f9ef4
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>