scripts: refactor fio test scripts
The iSCSI and NVMe-oF fio test scripts have a lot duplicated codes.
So create the fiocommon.py and initdevice.py file to have the codes
used for both. And also create iscsi_fio.py and nvmf_fio.py files
to test iSCSI and NVMe-oF with fio.
The devices list will be passed into iscsi_fio.py and nvmf_fio.py.

Change-Id: Ib8eb65f55f5975a4a7c9155c2b2f78d65736b317
Signed-off-by: Liang Yan <liang.z.yan@intel.com>