Update patch set 1
Patch Set 1:

I'm a bit confused as to why this is here and not part of https://review.gerrithub.io/#/c/386563/ and did see Karol's comment on that review pointing to this one.  Karol, maybe you can give me a quick high level overview of why these should be split in two versus being in the same patch?

Patch-set: 1
Label: Verified=0