Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Before recent changes in initialization of blobstore only umap was used as a way to assure no metadata was accidentally used in new blobstore.
I agree that since write_zeroes is used to assure that, replacing unmap for all blobstore oprations should not be needed.

Patch-set: 1