Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Can you create a unit test that shows the error in the original implementation, and shows how this patch fixes it?

A lot of time was spent on unit tests for this reset code - if there are gaps we should fill them as the code gets updated.

Patch-set: 2
Label: Code-Review=-1