Update patch set 2
Patch Set 2: Code-Review-1

Thank you for your proposal. The purpose of this patch looks good.

I think if we want to remove all dependency on the DPDK, we should consider how to refactor the constructor based on dpdk_mempool_create(), that is spdk_mobj_ctor and spdk_iscsi_sess_ctor.

I think you have already considered that but it will be helpful if you describe concretely the route to the goal anywhere even just a little bit.

Patch-set: 2