bdev: Unit test that resets don't complete early
Demonstrate a bug in the reset path where the reset
completes before the channels have all been notified.
This leaves the code open to race conditions where, based
on the reset completion a message could be passed to
a thread to issue I/O, but that thread may still be
blocked.

Change-Id: I84deeca40cc35497677367cb7aa933a4670f0333
Signed-off-by: Ben Walker <benjamin.walker@intel.com>