vtophys: remap vfio dma memory when necessary
VFIO requires at least one IOMMU group to be added to the
VFIO container before it possible to perform any IOMMU operations
on that container. [1] Without any groups added, VFIO_IOMMU_MAP_DMA
would always respond with errno 22 (Invalid argument).

Consider the following scenario:

1. start vhost, create an empty controller
2. start a vm, a warning is printed - vm memory fails to map
3. hotplug nvme
4. vm memory is not mapped for DMA, nvme dma will fail

Also, if the last IOMMU group is removed from a controller, it's IOMMU
mappings are lost as well. Consider another scenario:

1. start vhost app with a single nvme drive
2. start vm, memory is mapped successfully
3. hotremove nvme, vm memory will be unmapped
4. hotplug nvme
5. vm memory is not mapped for DMA, nvme dma will fail

This patch adds PCI device ref counter to our vfio manager and keeps
track of active vfio mappings. When refcount is increased to 1,
the vfio will remap all dma memory.

[1] https://www.kernel.org/doc/Documentation/vfio.txt
"On its own, the container provides little functionality, with all
but a couple version and extension query interfaces locked away.
The user needs to add a group into the container for the next level
of functionality.  [...] With a group (or groups) attached to a
container, the remaining ioctls become available, enabling access
to the VFIO IOMMU interfaces."

Change-Id: I744e07043dbe7ffd433fc95d604dad39647675f4
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>