Update patch set 4
Patch Set 4: Code-Review-1

(1 comment)

this seems to be a lot of complexity

could we instead add a function pointer in virtio_req?  then just call this function in bdev_virtio_poll

yes - it's another level of indirection but I really do not think it will be noticeable at all

this would make things far simpler - i'm concerned that when we want to do rescans at runtime, we'll need to add refcounts etc. to make sure we accurately revert to the faster I/O poll routine again - with the function pointer, that all becomes very simple and we don't have to unregister/re-register pollers

Patch-set: 4
Label: Code-Review=-1