Update patch set 4
Patch Set 4:

(1 comment)

> Patch Set 4: Code-Review-1
> 
> (1 comment)
> 
> this seems to be a lot of complexity
> 
> could we instead add a function pointer in virtio_req?  then just call this function in bdev_virtio_poll
> 
> yes - it's another level of indirection but I really do not think it will be noticeable at all

If we're about to do add another layer of indirection just for the scan, we might as well make bdev_virtio_scan_poll the default I/O poller. Will that do? Another abstraction layer here is not necessary just for this purpose  and I'd like to avoid adding it if it's not necessary.

> this would make things far simpler - i'm concerned that when we want to do rescans at runtime, we'll need to add refcounts etc. to make sure we accurately revert to the faster I/O poll routine again - with the function pointer, that all becomes very simple and we don't have to unregister/re-register pollers

refcounts? I think there can be only one scan at a time (per spdk process). This patch adds a safety check against scanning the device multiple times - bdev_virtio.c:1335.
Since all the scan I/O is done on a single thread it doesn't seem a problem to unregister/register a poller, it's all done synchronously.

The scan logic is about to be rewritten as well - probably as a state machine with states for both sending a request and receiving proper response. Would that do instead?

Patch-set: 4