Update patch set 37
Patch Set 37: Code-Review-1

This adds 2 minutes to the test run time on fedora-05.  That is a lot of time.

Is there any way we can fold this into an existing test, so that we do not have to start VMs just to test read-only functionality?  Maybe have two different vhost-blk devices - one read/write, the other read-only?

Patch-set: 37
Label: Code-Review=-1