Update patch set 14
Patch Set 14: Code-Review-1

My intention here was that the user could fill out an spdk_blob_opts structure and do everything necessary to add initial xattrs, set size and in the future add thin provisioning flag in the create call.  Then for snapshot/clone use cases we have a single metadata operation for the blob instead of one for create and another to update metadata.

I don't think Ben suggested removing the size from the options - just that it should be named something like num_clusters to make it clear the size was in terms of clusters (not bytes or pages).

Patch-set: 14
Label: Code-Review=-1